Moved `favicon` out of the middleware
ref no-issue

- The favicon middleware makes a comparison on every request to check if
it should serve the favicon. Adding the path prevents this comparison
and skips the function entirely on non-favicon requests. Moving it to
the `.get` avoids unnecessary comparisons on other http methos (POST,
DELETE, PUT, ...) and only serves it on GET requests.
