+++ ghost/core/core/frontend/web/middleware/index.js
@@ -4,7 +4,6 @@ module.exports = {
     frontendCaching: require('./frontend-caching'),
     handleImageSizes: require('./handle-image-sizes'),
     redirectGhostToAdmin: require('./redirect-ghost-to-admin'),
-    serveFavicon: require('./serve-favicon'),
     servePublicFile: require('./serve-public-file'),
     staticTheme: require('./static-theme')
 };
